+++ src/Processors/Transforms/FilterTransform.h
@@ -23,7 +23,7 @@ public:
     FilterTransform(
         const Block & header_, ExpressionActionsPtr expression_, String filter_column_name_,
         bool remove_filter_column_, bool on_totals_ = false, std::shared_ptr<std::atomic<size_t>> rows_filtered_ = nullptr,
-        std::optional<std::pair<size_t, String>> condition_ = std::nullopt);
+        std::optional<std::pair<UInt64, String>> condition_ = std::nullopt);
 
     static Block
     transformHeader(const Block & header, const ActionsDAG * expression, const String & filter_column_name, bool remove_filter_column);
@@ -48,7 +48,7 @@ private:
     std::shared_ptr<std::atomic<size_t>> rows_filtered;
 
     /// If set, we need to update the query condition cache at runtime for every processed chunk
-    std::optional<std::pair<size_t, String>> condition;
+    std::optional<std::pair<UInt64, String>> condition;
 
     std::shared_ptr<QueryConditionCache> query_condition_cache;
